+++ CHANGELOG.md
@@ -3,6 +3,11 @@ Added this change log after project was already under way.  Early changes are no
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [2.1] - 2019-02-10
+### Changed
+- added shortcut method `toJSON()` to `JSONSerializer` to output directly to string
+- added shortcut methods `parse()` to `JSONDeserializer` to parse directly from string
+
 ## [2.0] - 2019-02-10
 ### Changed
 - switched to Java 8; added handling of `java.time` classes and `Optional`